Character Selection Added
Adds character selection to the new game -- with create and load options with file dialog working. Right now this loads and saves text files -- does not update a persistent character object/singleton for example.

Note: this also requires a characterFiles folder in the user:// directory. Example character added to the _toArchive folder for now.