Adds Merged Game Demo Web Version
This is an add of the merged game demo for web with required cleanup and code changes for it to run in web. Currently the load character csv does not function, this is still in development for the web version.

To view served via itch.io: https://macdug.itch.io/bcirpg-game-base-test