Updates MVP Game Code for Player Singleton Updates
This update creates a merged MVP Game Code incorporating Luke's player updates. Also:

1. Makes modifications to the import character options, working from this set singleton
2. Adjusts game code to call print of character stats for the character panel output
3. Adds demo csv to the _userFiles

NOTE: csvs made in this process use an initial line with headers, and a following line with matching values. Code modifications assume this format and work (at this point) from hard-coded character values. We'll need to add the abstraction layer on top of this for variable game display.