Character Conversion Work CONT'D 5
Adds Falstaff csv test character for OpenD6 to the _userFiles. Adds functionality to the import to match the add character. Fixes bug in import, and fixes that bug in percentile character for both add and import as well. Finally, fixes character print in-game so percentile displays as well as variable front-end games.