Made the font size selection functional, I guess.
As far as I can tell, the only way to change font size is to replace the whole font, so I added a function to do that. It's in globalSaveInstance, which might not be the best place to put it. I run the function in the main menu and game scene right after the theme is loaded.