De-hardcoded theme selection
I chose to implement it in a slightly funky way, so please judge. The settings menu searches the assets/Themes folder for theme files and creates a dictionary with them (no duplicate names!). The file path of the selected theme is saved. I don't know how well this will play with a bundled version of the game.